server/requirements/requirements.txt
astroid==1.4.9
backports.functools-lru-cache==1.3
colorama==0.3.7
configparser==3.5.0
Django==1.10.4
django-braces==1.11.0
django-environ==0.4.1
django-oauth-toolkit==0.11.0
django-pyodbc-azure==1.10.4.0
django-rest-framework-social-oauth2==1.0.5
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
google-api-python-client==1.6.2
httplib2==0.10.3
isort==4.2.5
jsonpickle==0.9.4
lazy-object-proxy==1.2.2
mccabe==0.5.3
oauth2client==4.0.0
oauthlib==1.1.2
pyasn1==0.2.2
pyasn1-modules==0.0.8
PyJWT==1.4.2
pylint==1.6.4
pyodbc==4.0.1
python-openid==2.2.5
requests==2.13.0
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
uritemplate==3.0.0
wrapt==1.10.8
